Move apportion to protocol refinement
* More sensible than lambda schedule
