+++ tmd/fe/protocol_refinement.py
@@ -1,4 +1,5 @@
 # Copyright 2019-2025, Relay Therapeutics
+# Modifications Copyright 2026, Forrest York, Justin Gullingsrud
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,6 +14,9 @@
 # limitations under the License.
 
 from typing import Callable, TypeVar
+
+import numpy as np
+from numpy.typing import NDArray
 
 _T = TypeVar("_T")
 
@@ -61,3 +65,35 @@ def copy_and_insert(xs: list[_T], idx: int, x: _T) -> list[_T]:
     xs_ = xs.copy()
     xs_.insert(idx, x)
     return xs_
+
+
+def apportion(weights: NDArray, total: int) -> NDArray:
+    """Apportion `total` items proportionally to `weights` using the largest-remainder method.
+
+    Each entry receives at least floor(weight_i / sum * total) items; the remaining
+    items are assigned one each to the entries with the largest fractional remainders.
+
+    Parameters
+    ----------
+    weights : array of float, length n
+        Non-negative weights (must sum to > 0).
+    total : int
+        Total number of items to distribute (must be >= len(weights)).
+
+    Returns
+    -------
+    NDArray of int, length n
+        Integer allocation summing to `total`.
+    """
+    assert len(weights) > 0, "Must provide at least one weight"
+    assert total >= len(weights), "total must be greater than the number of weights"
+    assert np.all(weights >= 0.0), "All weights must greater than or equal to 0"
+    raw = weights / weights.sum() * total
+    floored = np.floor(raw).astype(int)
+    remainders = raw - floored
+    deficit = total - floored.sum()
+    if deficit > 0:
+        bonus_idxs = np.argsort(remainders)[-deficit:]
+        floored[bonus_idxs] += 1
+    assert floored.sum() == total, "Failed to apportion values"
+    return floored
